Renamed asset to image
