+++ src/components/opinion/cutout.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { css, SerializedStyles } from '@emotion/core';
 
 import { Contributor, isSingleContributor } from 'capi';
-import { transformUrl } from 'asset';
+import { transformUrl } from 'image';
 
 
 // ----- Constants ----- //
